split up slow replication tests
